Final linting, prettier, css and pathing changes
