+++ src/components/Navbar/ButtonForNavbar.tsx
@@ -1,26 +1,21 @@
+import React, { useState } from 'react';
 import Button from '@mui/material/Button';
 import Typography from '@mui/material/Typography';
-import React from 'react';
 
 interface ButtonProps {
   location: string;
   page: string;
 }
 
-const ButtonForNavbar = ({ page, location }: ButtonProps) => {
-  const [, setAnchorElNav] = React.useState<null | HTMLElement>(null);
+const ButtonForNavbar: React.FC<ButtonProps> = ({ page, location }) => {
+  const [, setAnchorElNav] = useState<null | HTMLElement>(null);
 
   const handleCloseNavMenu = () => {
     setAnchorElNav(null);
   };
 
-  const checkPage = () => {
-    if (location === '/' && page === 'Masterprojects') {
-      return true;
-    } else {
-      return false;
-    }
-  };
+  const checkPage = () => location === '/' && page === 'Masterprojects';
+
   return (
     <Button onClick={handleCloseNavMenu} sx={{ my: 2, color: '#3D7844', display: 'block' }}>
       {checkPage() ? (
