+++ src/components/Navbar/Layout.tsx
@@ -1,18 +1,18 @@
-import React, { ReactNode } from 'react';
+import React from 'react';
 import Navbar from './Navbar';
 import './Layout.css';
 
 interface LayoutProps {
-  children: ReactNode;
+  children: React.ReactNode;
 }
 
-function Layout({ children }: LayoutProps) {
+const Layout: React.FC<LayoutProps> = ({ children }) => {
   return (
     <div>
       <Navbar />
       <div className="content-below-navbar">{children}</div>
     </div>
   );
-}
+};
 
 export default Layout;
