+++ src/components/ProjectDetail/InformationBox.tsx
@@ -1,5 +1,5 @@
-import { Stack } from '@mui/material';
 import React from 'react';
+import { Stack } from '@mui/material';
 import { locations, studyFields } from '../../models/allowedValues';
 import CalendarTodayIcon from '@mui/icons-material/CalendarToday';
 import NotificationsNoneIcon from '@mui/icons-material/NotificationsNone';
@@ -13,7 +13,7 @@ interface InformationBoxProps {
   studyField: string;
 }
 
-const InformationBox = ({ deadline, location, duration, studyField }: InformationBoxProps) => {
+const InformationBox: React.FC<InformationBoxProps> = ({ deadline, location, duration, studyField }) => {
   return (
     <Stack
       spacing={2}
@@ -27,7 +27,7 @@ const InformationBox = ({ deadline, location, duration, studyField }: Informatio
     >
       <div style={{ fontSize: '20px' }}>
         <div className="projectDetailDate">
-          <LocationOnIcon style={{ color: '#3D7844' }}></LocationOnIcon>
+          <LocationOnIcon style={{ color: '#3D7844' }} />
           <div style={{ fontSize: '20px' }}>{locations[location as keyof typeof locations]}</div>
         </div>
       </div>
@@ -35,17 +35,16 @@ const InformationBox = ({ deadline, location, duration, studyField }: Informatio
       <div style={{ fontSize: '20px' }}>{studyFields[studyField as keyof typeof studyFields]}</div>
       <div style={{ fontSize: '20px' }}>
         <div className="projectDetailDate">
-          <CalendarTodayIcon style={{ color: '#3D7844' }}></CalendarTodayIcon>
+          <CalendarTodayIcon style={{ color: '#3D7844' }} />
           <div style={{ margin: '10px' }}>{duration}</div>
         </div>
       </div>
       <div style={{ fontSize: '20px' }}>
         <div style={{ fontSize: '20px' }}>
           <div className="projectDetailDate">
-            <NotificationsNoneIcon style={{ color: '#3D7844' }}></NotificationsNoneIcon>
+            <NotificationsNoneIcon style={{ color: '#3D7844' }} />
             <div style={{ margin: '10px' }}>
-              {' '}
-              Application deadline: {deadline.getDate()}/{deadline.getMonth()}/{deadline.getFullYear()}{' '}
+              Application deadline: {deadline.getDate()}/{deadline.getMonth() + 1}/{deadline.getFullYear()}
             </div>
           </div>
         </div>
@@ -53,4 +52,5 @@ const InformationBox = ({ deadline, location, duration, studyField }: Informatio
     </Stack>
   );
 };
+
 export default InformationBox;
