+++ src/components/UploadProjectForm/UploadProjectForm.tsx
@@ -1,11 +1,13 @@
-import React from 'react';
-import '../pages/uploadProject/uploadProject.css';
-import { Box, Button, MenuItem, Select, TextField } from '@mui/material';
-import { SetStateAction, useState } from 'react';
-import createExperienceReport from '../services/createExperienceReport';
-import { allowedLocations, allowedStudyFields } from '../models/allowedValues';
+import React, { SetStateAction, useState } from 'react';
+import './uploadProjectForm.css';
+import { Box, Button, MenuItem, Select, TextField, TextFieldProps } from '@mui/material';
+import { DatePicker } from '@mui/x-date-pickers';
+import createProject from '../../services/createProject';
+import { allowedLocations, allowedStudyFields } from '../../models/allowedValues';
 import { useNavigate } from 'react-router-dom';
-const UploadExperienceReport = () => {
+
+const UploadProjectForm: React.FC = () => {
+  const [value, setValue] = useState<string | null>(Date());
   const [studyField, setStudyField] = useState('study_field');
   const [location, setLocation] = useState('location');
   const navigate = useNavigate();
@@ -33,8 +35,8 @@ const UploadExperienceReport = () => {
       return;
     }
     const data = new FormData(event.currentTarget);
-    const year = parseInt(data.get('year') as string);
-    if (Number.isNaN(year)) {
+    const duration = parseInt(data.get('duration') as string);
+    if (Number.isNaN(duration)) {
       alert(`Duration must be a number`);
       return;
     }
@@ -51,34 +53,43 @@ const UploadExperienceReport = () => {
 
     try {
       const shortTitle = data.get('shortTitle') as string;
-      const experienceReportTitle = data.get('experienceReportTitle') as string;
+      const projectTitle = data.get('projectTitle') as string;
 
+      const dateStr = data.get('date') as string;
+      const dateParts = dateStr.split('/');
+      const year = parseInt(dateParts[2], 10);
+      const month = parseInt(dateParts[1], 10) - 1; // January is 0, so subtract 1
+      const day = parseInt(dateParts[0], 10);
+
+      const deadline = new Date(year, month, day);
       const description = data.get('description') as string;
       const summaryDescription = data.get('summaryDescription') as string;
-      createExperienceReport(
-        experienceReportTitle,
+
+      createProject(
+        projectTitle,
         shortTitle,
         studyField,
         location,
-        year,
+        deadline,
         description,
-        summaryDescription,
-        'hall'
+        duration,
+        summaryDescription
       );
-      console.log('Successfully upload');
+      console.log('Successfull upload');
     } catch (error) {
       console.log(error as string);
       return;
     }
     navigate('/');
   };
+
   return (
     <Box component="form" noValidate onSubmit={handleUpload} sx={{ margin: '0 auto', maxWidth: '60%' }}>
       <Box display={'flex'} sx={{ marginBottom: '1em', marginTop: '1em' }}>
         <TextField
           fullWidth
           id="shortTitle"
-          label="Experience report short title"
+          label="Project short title"
           name="shortTitle"
           sx={{ marginRight: '1em' }}
         />
@@ -100,6 +111,14 @@ const UploadExperienceReport = () => {
           <MenuItem value="water_and_sanitation">Water and sanitation</MenuItem>
         </Select>
       </Box>
+      <TextField
+        required
+        fullWidth
+        id="projectTitle"
+        label="Project Title"
+        name="projectTitle"
+        sx={{ marginBottom: '1em' }}
+      />
       <Box display={'flex'} sx={{ marginBottom: '1em' }}>
         <Select
           id="location"
@@ -116,16 +135,8 @@ const UploadExperienceReport = () => {
           <MenuItem value="south_america">South America</MenuItem>
           <MenuItem value="north_america">North America</MenuItem>
         </Select>
-        <TextField required fullWidth id="year" label="Year" name="year" sx={{ marginLeft: '1em' }} />
+        <TextField required fullWidth id="duration" label="Duration" name="duration" sx={{ marginLeft: '1em' }} />
       </Box>
-      <TextField
-        required
-        fullWidth
-        id="experienceReportTitle"
-        label="Experience report title"
-        name="experienceReportTitle"
-        sx={{ marginBottom: '1em' }}
-      />
       <TextField
         required
         fullWidth
@@ -143,6 +154,21 @@ const UploadExperienceReport = () => {
       />
 
       <Box display={'flex'} sx={{ marginBottom: '1em' }}>
+        <Box sx={{ marginRight: '1em', width: '100%' }}>
+          <div>Application deadline</div>
+          <DatePicker
+            disablePast
+            value={value}
+            onChange={(newValue) => {
+              setValue(newValue);
+            }}
+            renderInput={(params: JSX.IntrinsicAttributes & TextFieldProps) => (
+              <TextField id="date" name="date" InputLabelProps={{ shrink: false }} {...params} />
+            )}
+            inputFormat="DD/MM/YYYY"
+          ></DatePicker>
+        </Box>
+
         <Button type="submit" fullWidth variant="contained" sx={{ marginLeft: '1em' }}>
           Upload
         </Button>
@@ -151,4 +177,4 @@ const UploadExperienceReport = () => {
   );
 };
 
-export default UploadExperienceReport;
+export default UploadProjectForm;
