+++ src/pages/signupPage/SignupPage.tsx
@@ -1,6 +1,6 @@
 import { createTheme, ThemeProvider } from '@mui/material/styles';
 import './signupPage.css';
-import SignUpComponent from '../../components/Signup';
+import SignUpComponent from '../../components/Signup/Signup';
 import Logo from './../../images/logo.png';
 import ImageIcon from './../../images/loginBilde.png';
 import React from 'react';
